Add some support for FW upload to RAM
Two main things here
1. firmware image transfer from host to target is the upload
from host perspective, so rename /download/upload/g is in order.
2. It is sometimes desirable to upload the (untested) image RAM
rater than to flash. To do that, the option to choose the
programming script is necessary. By default it should be
"program" which will instruct OpenOCD to send image to flash,
but it may also be load_image which will send it to RAM,
or it shall be also possible to run custom tcl proc here,
that does any of the above and any additional necessary setup.
It may also be beneficial to provide custom reset command.

These two points are addressed in this CL:
- global find/replace of all the "download" references with "upload".
- 2 edit fields added to enter upload method name and reset sequence.

Signed-off-by: Alexey Polyudov <apolyudov@google.com>
Change-Id: I7e4f84ca5b79a4e2c17c9c3596170e9e74dcfef4